+++ supplier_api_tester/supplier_api_tester/v1/tests/__init__.py
@@ -51,7 +51,8 @@ BOOKING_TEST = (
     booking.test_incorrect_api_key,
     booking.test_booking_incorrect_reservation_id,
     booking.test_booking,
-    booking.test_cancellation
+    booking.test_cancellation,
+    booking.test_barcodes,
 )
 
 PRODUCT_CATALOG = (