+++ supplier_api_tester/supplier_api_tester/v1/tests/booking.py
@@ -1,6 +1,6 @@
 # If multiple checks can be done using the same response
 # then they should be done under a single test case.
-
+import base64
 from datetime import datetime
 import time
 
@@ -99,6 +99,7 @@ def test_booking_incorrect_reservation_id(api_url, api_key, product_id, timeslot
     check_api_error(raw_response, api_error, expected_error)
     return TestResult()
 
+
 @test_wrapper
 def test_booking(api_url, api_key, product_id, timeslots: bool, version=1):
     '''Booking tickets for at least 1 variant'''
@@ -136,6 +137,7 @@ def test_booking(api_url, api_key, product_id, timeslots: bool, version=1):
                 )
 
     return TestResult()
+
 
 @test_wrapper
 def test_cancellation(api_url, api_key, product_id, timeslots: bool, version=1):
@@ -239,3 +241,39 @@ def test_cancellation(api_url, api_key, product_id, timeslots: bool, version=1):
     check_api_error(raw_response, api_error, expected_error)
     return TestResult()
 
+
+@test_wrapper
+def test_barcodes(api_url, api_key, product_id, timeslots: bool, version=1):
+    """Barcodes match the expected barcode format."""
+    url = f'{api_url}/v{version}/products/{product_id}/reservation'
+    slot = get_reservation_slot(api_url, api_key, product_id, timeslots)
+
+    json_payload = get_payload_from_slot(slot, variant_quantity=2, min_quantity=3)
+    if timeslots:
+        json_payload['timeslot'] = slot.start
+    raw_response, response = client(url, api_key, method=requests.post, json_payload=json_payload)
+    reservation = get_reservation(raw_response, response)
+
+    url = f'{api_url}/v{version}/booking'
+    raw_response, response = client(url, api_key, method=requests.post, json_payload={
+        'reservation_id': reservation.reservation_id,
+        'order_reference': reference_id(),
+    })
+    booking = get_booking(raw_response, response)
+    barcodes: list[str] = []
+
+    if booking.barcode_format.lower() in ['aztec-bytes', 'pdf']:
+        if booking.barcode_position == 'order':
+            barcodes.append(booking.barcode)
+        else:
+            for _, tickets in booking.tickets.items():
+                barcodes.extend(tickets)
+        for barcode in barcodes:
+            try:
+                base64.b64decode(barcode).decode('utf-8')
+            except Exception:
+                return TestResult(
+                    status=2,
+                    message=f'Expected base64-encoded string for barcode type {booking.barcode_format}: {barcode}',
+                )
+    return TestResult()