+++ supplier_api_tester/supplier_api_tester/v2/tests/__init__.py
@@ -34,6 +34,7 @@ BOOKING_TEST = (
     booking.test_booking_incorrect_reservation_id,
     booking.test_booking,
     booking.test_cancellation,
+    booking.test_barcodes,
 )
 
 PRODUCT_CATALOG = (